requirements.txt
amqp==5.2.0
asgiref==3.7.2
billiard==4.2.0
celery==5.4.0
certifi==2024.6.2
channels==4.1.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
diff-match-patch==20230430
Django==5.0.2
django-environ==0.11.2
django-import-export==4.1.1
idna==3.7
kombu==5.3.7
mysqlclient==2.2.4
pillow==10.3.0
prompt_toolkit==3.0.45
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
requests==2.32.3
six==1.16.0
sqlparse==0.4.4
stripe==9.11.0
tablib==3.5.0
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
